soft mmu: some work on warning and a few small refactoring, especially around variable argument functions.

+++ source/emulation/cpu/normal/normal_other.h
@@ -175,7 +175,7 @@ void OPCALL normal_retn32(CPU* cpu, DecodedOp* op) {
 }
 void OPCALL normal_invalid(CPU* cpu, DecodedOp* op) {
     START_OP(cpu, op);
-    kpanic("Invalid instruction %x\n", op->inst);
+    kpanic("Invalid instruction %x\n", static_cast<int>(op->inst));
 }
 void OPCALL normal_int80(CPU* cpu, DecodedOp* op) {
     START_OP(cpu, op);
